requirements.txt
pandas==2.0.3
requests==2.31.0
apache-airflow-providers-postgres==5.10.2
apache-airflow-providers-amazon==8.0.0
boto3==1.26.138
Pillow==9.5.0

apache-airflow[amazon,postgres]==2.8.1
